small fixes, moved init back
